clippy

+++ zi/src/language_service/lstypes.rs
@@ -190,20 +190,20 @@ pub(crate) trait TextExt {
 
 impl<T: Text> TextExt for T {
     fn decode_point(&self, EncodedPoint { point, encoding }: EncodedPoint) -> Option<Point> {
-        if point.line() as usize > self.len_lines() {
+        if point.line() > self.len_lines() {
             return None;
         }
 
         match encoding {
             PositionEncoding::Utf8 => Some(Point::new(point.line(), point.col())),
             PositionEncoding::Utf16 => {
-                let line_start_byte = self.line_to_byte(point.line() as usize);
+                let line_start_byte = self.line_to_byte(point.line());
                 let line_start_cu = self.byte_to_utf16_cu(line_start_byte);
-                if line_start_cu + point.col() as usize > self.len_utf16_cu() {
+                if line_start_cu + point.col() > self.len_utf16_cu() {
                     return None;
                 }
 
-                let byte = self.utf16_cu_to_byte(line_start_cu + point.col() as usize);
+                let byte = self.utf16_cu_to_byte(line_start_cu + point.col());
                 Some(self.byte_to_point(byte))
             }
         }
